requirements.txt
torch==2.4.0
torchvision==0.19.0
numpy==1.26.0
pandas==2.2.2
matplotlib==3.9.2
hydra-core==1.3.0